% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_tr_count.R
\name{add_tr_count}
\alias{add_tr_count}
\title{Collapse data while adding treatment count variable.}
\usage{
add_tr_count(collapsed_data, tr_group_name, count_col_name)
}
\arguments{
\item{collapsed_data}{Data that has been aggregated and includes a treatment
group variable as well as a count variable on the treatment group level
(data.table)}

\item{tr_group_name}{The name of the column in \code{collapsed_data} that
contains the names of the corresponding treatment groups (character)}

\item{count_col_name}{The name of the column in \code{collapsed_data} that
contains the treatment group-level observation counts (character)}
}
\description{
\code{add_tr_count} takes collapsed data and adds a treatment
    group variable that reflects the number of observations in that treatment
    bin. If the counts are unequal within treatment bins, the function gives
    a warning and adds an asterisk (*) after the count variable, which now
    represents the minimum of all the counts within each treatment group bin.
}
