% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/balance_panel_data.R
\name{balance_panel_data}
\alias{balance_panel_data}
\title{Balance panel data.}
\usage{
balance_panel_data(panel_data, n_periods, panel_unit)
}
\arguments{
\item{panel_data}{The data to be balanced (data.table).}

\item{n_periods}{The number of time periods in the dataset (numeric).}

\item{panel_unit}{The name of the panel unit (character).}
}
\description{
This function takes a possibly unbalanced panel and returns
    a balanced panel over a given time period.
}
