% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_scanner_data.R
\name{combine_scanner_data}
\alias{combine_scanner_data}
\title{Function to combine Nielsen scanner data.}
\usage{
combine_scanner_data(folder = NULL, file_tail, file_type, years,
  filters = NULL, store_info_folder = folder,
  store_info_prefix = "stores_", store_info_filetype = file_type,
  select_modules = F, modules_data = NULL)
}
\arguments{
\item{folder}{The folder in which the scanner data and the year-level store
information are contained (if they aren't in the working directory) (character).}

\item{file_tail}{The name of the scanner data files (that follows the year). For
example, if the 2010 file was called \code{2010_scanner_data.dta}, the
\code{file_stem} argument would be \code{"_scanner_data"}. (character)}

\item{file_type}{The type of the file. Currently works for either \code{.dta},
\code{.csv}, and \code{.txt} files. (character)}

\item{years}{A vector of years for which the scanner data should be returned
(numeric).}

\item{filters}{A string containing the filtering conditions (e.g., if I wanted to keep
stores in fips_state == 4 and fips_county == 18, it would be
\code{"fips_state == 4 & fips_county == 18"}) (character)}

\item{store_info_folder}{The folder containing the store information files,
if different than \code{folder}. (character)}

\item{store_info_prefix}{The prefix for the store information files. (character)}

\item{store_info_filetype}{The filetype of the store information files, if
different than \code{file_type}. (character)}

\item{select_modules}{Do you want to specify specific product modules to
keep? (logical)}

\item{modules_data}{If \code{select_modules = TRUE}, the dataset containing
the modules to be kept (data.table)}
}
\description{
This function combines year-level Nielsen data cross years,
    keeping only a subset of stores specified by the user. The user also
    specifies for which years he would like the data.
}
