% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_pseudo_control.R
\name{create_pseudo_control}
\alias{create_pseudo_control}
\title{Create pseudo-control observations for event-study plots}
\usage{
create_pseudo_control(es_data, original_data, control_counties, time_var)
}
\arguments{
\item{es_data}{The dataset of event-study like data (data.table)}

\item{original_data}{The original data (that still includes the control
group data) (data.table)}

\item{control_counties}{A dataset of identifying information (state and county
fips codes) for control counties (data.table)}

\item{time_var}{The type of time variable (should be \code{"quarter"} or \code{"month"})}
}
\value{
A modified version of \code{es_data} that has matched each treated
    observation to a control observation mean and added those "matched"
    observations as new observations in the dataset.
}
\description{
\code{create_pseudo_control} creates a set of control observations
    for event study graphs based on the control average at the corresponding
    calendar time.
}
