% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/es_price_application.R
\name{es_price_application}
\alias{es_price_application}
\title{Make event study graphs for prices (temporary file)}
\usage{
es_price_application(price_data, treatment_data_path,
  county_pop_data = NULL, county_sales_weights = NULL, weighting_var,
  price_var, resid_type = NULL, w_tax, fig_outfile = NULL)
}
\description{
Make event study graphs for prices (temporary file)
}
