% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_tax_rates.R
\name{identify_exemptions}
\alias{identify_exemptions}
\title{Merge tax rates (county-product level)}
\usage{
identify_exemptions(sales_data, county_monthly_tax_data,
  module_exemptions_path = "/project2/igaarder/Data/modules_exemptions_long.csv")
}
\arguments{
\item{sales_data}{The county-product level dataset, including quantities and/
or sales information (but not tax rates) (data.table)}

\item{county_monthly_tax_data}{A dataset of county-by-month tax rates
(data.table)}

\item{module_exemptions_path}{The path to the .csv file containing a long
dataset of state-product level exemptions/special tax rates. The default
is \code{"/project2/igaarder/Data/modules_exemptions_long.csv"}. Don't
change this unless you know what you are doing (character)}
}
\description{
\code{merge_tax_rates} merges the applicable tax rates to a
    county-product level dataset, including exemptions or reduced rates.
}
