% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_best_selling_products.R
\name{keep_best_selling_products}
\alias{keep_best_selling_products}
\title{Keep only best-selling product modules}
\usage{
keep_best_selling_products(all_data, module_name_ad, products_data,
  module_name_pd)
}
\arguments{
\item{all_data}{The data to be filtered, keeping only the best-selling
products found in \code{products_data}. Must contain the
\code{module_name_ad} column (data.table)}

\item{module_name_ad}{The name of the column in \code{all_data}
containing the identifying module names or numbers (character)}

\item{products_data}{A table of the best-selling products to keep. Must
contain the \code{module_name_pd} column (data.table)}

\item{module_name_pd}{The name of the column in \code{products_data}
containing the identifying module names or numbers (character)}
}
\description{
Filter product modules to keep only those found on the list
    of best-selling modules.
}
