% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_fixed_weights.R
\name{make_fixed_weights}
\alias{make_fixed_weights}
\title{Create weights based on a fixed time period}
\usage{
make_fixed_weights(panel_data, weight_time, weight_var, panel_unit_vars)
}
\arguments{
\item{panel_data}{The data for which weights are to be calculated (data.table)}

\item{weight_time}{The time period(s) that correspond to the time period of
the weights, as a list. For example, if the time period were
\code{year == 2008 & month == 1}, it would be correct to have
\code{weight_time = list(year = 2008, month = 1)} (list)}

\item{weight_var}{The name of the variable in \code{panel_data} that will
be used to create the weight (character)}

\item{panel_unit_vars}{The name(s) of the variables in \code{panel_data} that
uniquely identify the panel units for which the weights are to be
calculated (character)}
}
\description{
\code{make_fixed_weights} creates a weight for panel units
    based on a specific time period in the sample.
}
