% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_treatment.R
\name{merge_treatment}
\alias{merge_treatment}
\title{Merge treatment onto county-level sales data.}
\usage{
merge_treatment(original_data, treatment_data_path, time = "calendar",
  merge_by)
}
\arguments{
\item{original_data}{The panel data (without treatment status) (data.table)}

\item{treatment_data_path}{The path to the treatment data (character)}

\item{time}{Either \code{"calendar"} or \code{"event"} (character)}

\item{merge_by}{A vector of the names of the units to merge
    \code{original_data} and the treatment data by (character)

allow by to be either "calendar" or "event"
merge_by = c("fips_county", "fips_state")}
}
\description{
This function takes a county-level dataset and merges another
    dataset with treatment information onto the first dataset.
}
