% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/months_to_quarters.R
\name{months_to_quarters}
\alias{months_to_quarters}
\title{Aggregate monthly data to quarterly data}
\usage{
months_to_quarters(monthly_data, month_var, collapse_by, collapse_var,
  weights = F)
}
\arguments{
\item{monthly_data}{Dataset on month-level (data.table)}

\item{month_var}{The name of the month variable in \code{monthly_data} (character)}

\item{collapse_by}{The name(s) of the variable(s) that help define the
collapse (in addition to the seasons). Ideally, these variables will be
unique on the quarter-level (character)}

\item{collapse_var}{The name of the variable to be summed up over the seasons (character)}

\item{weights}{Not currently in use.}
}
\description{
\code{months_to_quarters} takes a monthly dataset and sums up
    a given variable to the quarterly level.
}
