% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalize_prices.R
\name{normalize_price}
\alias{normalize_price}
\title{Normalize prices}
\usage{
normalize_price(price_data, time_type, base_time, price_var,
  new_price_var = "normalized_price")
}
\arguments{
\item{price_data}{A dataset that includes prices to be normalized as well as
times of each price (data.table)}

\item{time_type}{Either \code{"calendar"} or \code{"event"}. The type of time
variable used in the data (character)}

\item{base_time}{The time of the base price. If \code{time_type = "event"},
\code{base_time} should be a single integer corresponding to the event time.
If \code{time_type = "calendar"}, \code{base_time} should be a vector of
two integers, the first corresponding to the year, the second corresponding
to the month (numeric)}

\item{price_var}{The name of the original price variable found in
\code{price_data} (character)}

\item{new_price_var}{The desired name of the new, normalized price variable
(default is just \code{"normalized_price"})}
}
\description{
\code{normalize_price} takes a dataset with prices and dates
    or event times and normalizes, using a given date/event time as the
    base.
}
