% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/one_cty_one_prod.R
\name{one_cty_one_prod}
\alias{one_cty_one_prod}
\title{Analyze outcome for one county, one product}
\usage{
one_cty_one_prod(product_data, fips_state, fips_county,
  product_module_code = NULL, upc = NULL, product_name,
  month_of_reform, year_of_reform, county_monthly_tax_data,
  rm_month_effects = FALSE,
  module_exemptions_path = "/project2/igaarder/Data/modules_exemptions_long.csv",
  balance_panel = T)
}
\arguments{
\item{product_data}{Dataset to be analyzed (data.table). Should have columns
with \code{fips_county}, \code{fips_state}, \code{product_module_code}}

\item{fips_state}{The state fips code of the desired state (integer)}

\item{fips_county}{The county fips code of the desired county (integer)}

\item{product_module_code}{The product module code of the desired product
(integer)}

\item{upc}{The name of the UPC code (if product is given on UPC level) (integer).
\code{upc} may be a list if you would like to average over multiple
UPCs (if, for example, a brand replaces a product with another, similar
product)}

\item{product_name}{The name of the product (for graph labeling) (character)}

\item{month_of_reform}{The month of the tax reform (integer)}

\item{year_of_reform}{The year of the tax reform (integer)}

\item{county_monthly_tax_data}{Dataset with tax rates (data.table)}

\item{rm_month_effects}{Remove month seasonality? (logical)}

\item{module_exemptions_path}{The path to the .csv file containing a long
dataset of state-product level exemptions/special tax rates. The default
is "/project2/igaarder/Data/modules_exemptions_long.csv"}

\item{balance_panel}{Force the panel to be balanced? (logical)}
}
\value{
A list including a \code{price_graph}, a \code{sales_graph}, and the
    data collapsed to event-time level (\code{})
}
\description{
\code{one_cty_one_prod} analyzes the given data for a single
    county for a single product type. The user must provide identifying
    information for the county and for either a product module code or a UPC
    (but not both).
}
