% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/price_application.R
\name{price_application}
\alias{price_application}
\title{Make event study graphs for prices (temporary file)}
\usage{
price_application(price_data, treatment_data_path, time, weighting_var,
  pretax_var, posttax_var, w_tax, fig_outfile = NULL)
}
\description{
Make event study graphs for prices (temporary file)
}
