% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/residualize_outcome.R
\name{residualize_outcome}
\alias{residualize_outcome}
\title{Function to residualize out covariates from a given dataset.}
\usage{
residualize_outcome(input_data, outcome_var, discrete_vars,
  continuous_vars, weight_var = NULL)
}
\arguments{
\item{input_data}{The data containing the outcome and covariates (data.table).}

\item{outcome_var}{The name of the outcome variable (character).}

\item{discrete_vars}{A list of discrete covariates (character).}

\item{continuous_vars}{A list of continuous variables (character).}

\item{weight_var}{(optional) The name of the weight variable, if the
regression is to be weighted (character).}
}
\description{
This function regresses an outcome on a given set of covariates and
 obtains and stores the residual for each observation.
}
