% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sales_application.R
\name{sales_application}
\alias{sales_application}
\title{Make event study graphs for sales (temporary file)}
\usage{
sales_application(sales_data, treatment_data_path, time,
  fig_outfile = NULL)
}
\description{
Make event study graphs for sales (temporary file)
}
